v1.0.9 and v1.0.10
Fixed an issue wherein pressing a button to open a menu (e.g. by interacting with a world object) also counted as pressing that button in the newly-opened menu. Fixed a crash in the MapMenu code caused by Bethesda failing to clear a pointer when redrawing the world map icons.