v1.0.18
Fixed a mistake in XXNAlchemyMenu::IngredientIsCompatible, wherein we failed to actually move to the next list element when iterating over an ingredient's linked list of effects. This should address issues where ingredients were properly listed in the menu, but were not selectable if the only matching effects elsewhere in the inventory were those after their first.