Zoom fix fix
Fixed a rare bug that could occur if multiple images in a hierarchy had their zoom traits change: some of them would append NiNodes directly to the center of the screen, possibly because the changing ancestor tiles didn't have NiNodes ready yet. This bug only occurred with DarN UI: the statbar fills in HUDMainMenu were affected, because they and their ancestors (the statbar backgrounds) both zoomed on the same trait, which itself changed whenever you entered and exited the Big Four menus.

The fix is to avoid blowing away image NiNodes; we don't need to use the "Text" changeflag, because the "Size" changeflag is enough to update the tiles per their new "zoom" values.