Add files via upload
Fixed generation of a missing NorthernUI.ctrl.txt file (and added code to regenerate it on startup, for people with affected versions). Added INI settings for PlayStation button icons/names in the UI, and for disabling the D-Pad as a menu navigation tool. Lowered the default Run Sensitivity from 95 to 92.