v1.2.2 work in progress
Fixed: Spellmaking incorrectly tells you that you aren't skilled enough to cast your spells. I mistook the spell requirement *display* for a spell requirement *warning* during the design stage. The DLL now supplies the SpellMakingMenu with the UI traits needed to check whether you can cast the new spell, so we can show the requirement along with a conditional warning.

HUDInfoMenu now uses reticle verbs defined in Oblivion.esm game-settings, though we still supplement those with custom strings. These game-settings were originally intended for the Xbox version, and though they're not used in the PC build, they're still loaded into an array whose indices exactly match the HUD reticle icon enum (offset by one).

The horiz_floating_scroll scrollbar now disables interaction on its buttons when the scrollbar is invisible; coupled with fixes in the EnchantmentSetting menu, this fixes issues with picking a lock level when making Lock or Open spells.

A patch to HUDMainMenu now makes it possible to rotate the player model seen in the Big Four menus using the gamepad. Rotation is mapped to whichever stick is not used for menu navigation.

NorthernUIAway: HUDInfoMenu now properly shows a button icon when a gamepad is plugged in. It should also show the reticle verb, but I haven't tested that yet.

NorthernUIAway: InventoryMenu should now consistently show the A-button mapping when a gamepad is plugged in.