v1.2.2 work-in-progress
Fixed the misidentified function EffectItemList::IsPoison (formerly listed as "HasNonPoisonousEffects"), and fixed a pertinent bug in XXNAlchemyMenu.

Patched InventoryMenu to restore Xbox functionality: showing the proper verb for any given item (and for sorting the inventory) as the label of the A-button. NorthernUI XML has been updated; NorthernUIAway has not been updated or tested yet.