v1.2.2 official release
The UI player model can no longer be rotated when MapMenu is open. This is due to a design issue: the vanilla menu uses the menu-nav stick to switch subpages in MapMeun, and the other stick to pan the map, so panning the map would've also uncontrollably rotated the player model and honestly that's just distracting.

NorthernUIAway (data-vanilla) content adjusted to make use of some of the original Xbox strings that we're now supplying to the menus. Fixed the Xbox "hint" prefab, used to display button mappings, to be visible by default if a gamepad is plugged in.