Movement and camera patches
Started work on adding new camera modes to Oblivion. In the "Oblivion Standard" mode, the camera behaves consistently with the vanilla game: the facing directions of the camera and the player-character are kept synchronized, and all movement is actor-relative. In the "Skyrim Standard" mode, the camera behaves consistently with Skyrim: you can turn the third-person camera independently of the character whenever you aren't moving and don't have a weapon drawn. In the "Free Movement" mode, the camera and the player are fully desynchronized: the player always rotates to face their movement direction, and the camera can be turned freely. The default mode is "Oblivion Standard."

Removed exploratory code that had been used to debug a cyclical reference issue in menu XML.