Movement and camera work
XXNOptionsMenu can now be used to configure the Enhanced Movement and Enhanced Camera patches.

The camera patch now has documentation comments, and it guards against out-of-bounds values from the INI file.

Fixed multiple mistakes that broke first-person view: the SnapForFirstPerson hook was unnecessary and actually made camera rotation impossible; the rest of the camera patches needed to specifically ignore first-person view, because for correct first-person view behavior, the look stick and mouse must rotate the actor and the camera must synch with the actor.

The FinalizeCamera hook was actually a PrepCamera hook and has been renamed accordingly. (It runs before the other hooks, not after. During initial development I was looking for anything that accessed the camera floats, rather than reading the game's code sequentially.)

Removed the CameraRelativeMovement hook; it was rendered superfluous by the movement flags hook, and if enabled alongside the flags hook, it actually broke the movement behavior completely.

The camera patch now hooks the code that updates the HUD reticle, so that when using a non-Oblivion camera behavior, the reticle target is what *you* are looking at rather than what *your character* is looking at.

Attempted to hook the POV switch: the goal is to make it so that immediately upon entering first-person view, the player is snapped to the third-person camera's facing direction rather than having the camera snap to the actor direction. The hook doesn't work; I evidently need to hook elsewhere.

Identified where the game stores the reticle target: it's on InterfaceManager, with a non-member getter.