Movement and camera work
Fixed a circular reference in the prefab used by the "bolt-on" description shown for XXNOptionsMenu, which in turn made it possible to dynamically resize the "bolt-on" to fit the text.

The "Free" camera now forces the player to face forward if they have a ranged weapon or ranged spell drawn. The "Skyrim" camera now properly points the player camera-forward on their first frame of movement. Improved documentation comments for the enhanced movement and camera patches.

Decoded more of HUDMainMenu.