Movement and camera fixes
Critical bug fix: the Enhanced Movement patch no longer does anything to the player's movement when the player is dead. This fixes an issue that can cause the game to throw up a loading screen and softlock on it immediately after death (because for some reason, the game tries to move the player a zero distance upon death, and the movement math breaks when the player dies; we end up launching the player across Cyrodiil and triggering a load screen that conflicts with the death handling).

The alternate camera behaviors no longer influence camera control when the player is mounted on a horse.

We now snap the camera to the player-character on the first frame that they spend in a new worldspace or interior cell.

Fixed a regression in part of the Skyrim camera behavior (if you turn the camera apart from the player and then move, the player should snap to the camera; this wasn't happening) stemming from a function being called at the wrong time.

Added courtesy NOPs to the camera patches.

Added an extended console command, "!xxn forceCrashOnNextLoadScreen", which forces a crash on the next load screen. I had originally used this to help investigate the enhanced movement issue, by having my crash logger (in a separate DLL) dump the call stack on the next loading screen. It turned out to be totally unhelpful (I found the issue's cause after noticing that the camera would veer off into the sky for a couple frames before the softlock), so I'll probably remove it.