Camera inertia fix
Oblivion applies inertia to the third-person camera. However, this effect has a bug that sometimes causes the camera to swing wildly after you've passed through a load door. It almost looks intentional -- as if Bethesda wanted to show off the location you were stepping into -- but it's disorienting, it doesn't synch with movement, and it absolutely is a bug caused by the sudden change in your angle as you pass through the door. (Load doors are teleporters. It may look like your angle hasn't changed -- you're facing into the building when you enter, and you're facing into the building after you enter -- but the true angle can have changed.)

Added the ability to control whether we fix this (since again, it can resemble a feature), and the ability to wholly disable camera inertia as well.