Improved localization support
Added an xxnLocalization() tile, similar to NorthernUI() and xxnStrings(), which can be used to hold localization-specific menu configuration settings, e.g. settings that can make menus wider to accommodate languages with longer strings. The following menus offer hooks via this tile: XXNAlchemyMenu, XXNOptionsMenu, AudioMenu, ControlsMenu, EffectSettingMenu, GameplayMenu, HUDMainMenu, PauseMenu, StatsMenu, VideoMenu, and the "Big Four" menus generally.

XXNHUDClockMenu now stores its format strings in the xxnStrings() tile. The menu class no longer crashes when a format string is left undefined.

XXNOptionsMenu no longer crashes if one of the enumpicker value strings is left unspecified.

The label for the zoom control in MapMenu is now stored in xxnStrings().

When the XInput patch is disabled, the player is able to switch what joysticks are used for moving versus looking around via the ControlsMenu. The label we show for that is now in xxnStrings().

The bottom bar prefab now takes the Y-offset of its containing menu into account. Some menus had to be positioned at the -1 Y-coordinate due to texture shenanigans causing visible seams at the top edge of the screen; this change to the bottom bar should prevent a seam from appearing at the bottom of the screen.

Reorganized the font files. Font files now use a clear and consistent filename convention, and a translator can select which set of files to use (ANSI or Cyrillic) using the xxnLocalization() tile instead of having to rename files to overwrite each other.

Reverse-engineering: Improved our understanding of the player input/movement handler; we now know how the game gives skill experience for Athletics and Sneak.