Fixes
The 360 movement patch is on by default.

Made minor accommodations for WalkBlessed. These are not sufficient to fully address incompatibilities, and if you have it installed alongside NorthernUI, it's still possible to get yourself in a situation where you can't operate the camera. To explain the problem simply, WalkBlessed's "vanity mode" tweaks are also trying to implement a modern camera; WalkBlessed's implementation is the easiest possible working implementation, whereas my implementation is the *correct* implementation, and so the two unavoidably conflict. There's no fix short of intentionally seeking out every WalkBlessed patch site and writing code to revert all of its hooks (which as a side-effect would prevent ANYONE from using those patch sites -- not good).