Fixes
Fixed one of the button prefabs for NorthernUIAway: text is now properly centered when a gamepad isn't plugged in.

Removed redundant debugging code in Exploratory; turns out, one of the patch files had an older debugger for the same functionality, so I just swapped some stuff around.

Fixed a mismatched brace in Selectors.cpp that caused the xxnStrings() and NorthernUI() tiles to silently fail to load if xxnLocalization() failed to load.

Fixed an engine bug that seemed to have negative impacts on Tile::Value::AppendSrcOperator.