v1.3.2
Critical fix to tile-trait cyclical reference dump: it's now locked to an INI setting and disabled by default, since it can apparently generate multiple gigabytes of log spam.

The logging works by taking every single step in a cyclical reference loop, and logging all inbound references. This is incredibly dumb and I don't know how I didn't see the problem with it before now. What we SHOULD do instead is log only the cyclical reference loop itself, and THEN start at the end of the loop and traverse (and log) OUTBOUND references until we get back to the start of the loop. Or, uh, something like that.