Research into BSA handling
Researched how the game loads and manages BSAs, how the game "invalidates" files in BSAs, and how the game searches the loaded BSAs to locate a desired file. Findings of note include:

We've mapped out most of the data structure for a loaded BSA file. Some of its contents directly correspond to the file layout as described by UESP's documentation.

We've located the CompareFileTime call that the game uses to decide whether a loose file should override a BSA.

We've located the subroutine that the game uses to look up files from the loaded BSAs. Investigation indicates that BSAs cannot intentionally override each other's files: the game normally iterates over BSAs in load-order order, but if it finds a file in a given BSA, then that BSA is remembered; the next lookup will check that BSA first, out of order.

We've located the subroutine (LoadBSAFile) that loads a BSA file and adds it to the list of loaded BSA files (g_archiveList).

We've located the two NiTArray<BSHash>s used for ArchiveInvalidation.