Cursor work, to be reverted
Forgot to include the SleepWaitMenu class def in the last push.

Uploading some cursor work so I can revert it without losing it. I wanted to either: control the position that the cursor appears at when you enter MenuMode, using traits on the menu's root tile; or prevent the cursor from becoming fully opaque again when you enter MenuMode. Goal is to deal with user annoyances stemming from the cursor centering in the screen and being forced to full visibility when you enter dialogue with any NPC. However, it seems that that happening is the result of side-effects of side-effects of side-effects rather than any intentional action on Bethesda's part:

a) All vanilla code that intentionally changes the mouse cursor alpha has been accounted for. It's pretty much only changed by LoadingMenu and LockPickMenu.

b) All vanilla code that intentionally sets the cursor's position has been accounted for. Aside from InterfaceManager::UpdateCursorState, the cursor is pretty much only moved by LockPickMenu and the game's WndProc.

c) The code that forces the cursor to be visible in InterfaceManager::Update and InterfaceManager::UpdateCursorState isn't responsible for the cursor returning to full alpha; no-oping those calls did nothing.

The cursor is at offset 0x1C in InterfaceManager. I literally searched the executable for every occurrence of that constant and looked at every single occurrence of the form [NonESPRegister + 0x1C] within a square mile of all known InterfaceManager functions. I can say with 99.99% certainty that all direct access to the UI cursor has been accounted for and I can find NOTHING that intentionally -- barring Bethesda "intending" to rely on side-effects of side-effects -- resets the cursor to the center of the screen only when the first menu is opened and never on any other occasion.

The whole thing is a minor annoyance at best, and I'm sorry but Bethesda's decade-plus-old mouse cursor code is way too spaghetti (whether as originally written or as compiled) for this to be worth the effort to fix right now.