Cursor progress
Turns out, Bethesda's code wasn't spaghetti; I literally just overlooked something that turned out to be relevant. ShowLoadingMenu sets the cursor alpha to zero, and LoadingMenu's destructor forces it back to max; because the cursor alpha is at full outside of our control, when you next open any menu (including DialogMenu) the cursor starts off as initially visible. So yeah, that's not on Bethesda and I'll take the L there.

Patch is mostly functional now. There's still one minor bug: the "when all menus are closed" hook isn't working, so if you move the cursor, exit dialogue, and then open any menu any amount of time later, the cursor will still be visible. (Goal is to have the cursor always be initially hidden if you have a gamepad plugged in, such that it only appears if you move the mouse and auto-hides if you leave the mouse still for five seconds.)