Fixed a critical bug in the MenuQue compatibility code
Certain code patches can only apply once we've dealt with MenuQue, but the compatibility code wasn't giving them the go-ahead in cases where MenuQue was not detected. This is basically a showstopper bug and it's been present since v1.3.1.

Made preparations for adding script commands to work with the gamepad; these were planned for the next update but we need to get NorthernUI shippable ASAP due to the MenuQue issue.