Script command progress
OBSE plug-ins that define script commands must use nullptr as the parse function, because if a parse function is specified, it will be called by the Construction Set; the Cmd_Default_Parse function is only intended for use in-game and just calls a vanilla subroutine directly, so it crashes the CS.

Started work on a test script for the new commands. Not sure how I want it to work yet.

The DLL no longer sets up the OBSE messaging interface or its own localization data when loaded by the Construction Set.