String operand work
Redid how we handle string operands.

There’s only one way to implement this in a direct and straightforward manner: you patch Tile::UpdateTemplatedChildren to handle const operators differently. I’d already been thinking about doing it, but my research into MenuQue for compatibility purposes has revealed that MenuQue beat me to exactly the ideas I had, and that it implements those ideas in *the* only straightforward manner.

I have anxiety and I don't want to have to potentially explain the ins-and-outs of Oblivion's parser to the Nexus moderation team to prove I'm not being a plagiarist, so instead of doing this directly, I'm doing it indirectly. We now treat const operators with string operands as a totally different parse-code from normal const operators. As a bonus, we no longer need a conditional patch (i.e. patching one site only if MenuQue doesn’t already have it), which is good – if we always patch a given site, then a mod author wishing to analyze that site will always know that we patch it, at least if we’re installed.