SleepWaitMenu work
Investigating a bug that causes the game to softlock if you move the mouse cursor while SleepWaitMenu is ticking down and then press B to stop waiting early.

Our SleepWaitMenu format string patch only covered when the menu is shown, and not when the menu is ticking down; updated code here should address that.

Our operator patch now checks for a UI engine warning that can stem from a SRC operator not being a self-closing XML element; we log an additional message so that the warning is intelligible.