SleepWaitMenu fixed; investigating LockPickMenu
The SleepWaitMenu softlock has been explained and fixed.

Our gamepad support for LockPickMenu is inconsistent with the vanilla game: we handle the A button when it is released, but Oblivion handles the left mouse button when it is pressed. Patching the game to handle A on release hasn't worked: the game plays the "a tumbler has moved" sound but nothing else happens. Currently I'm investigating tumbler state values so I can make sense of the code surrounding that sound.