LockPickMenu research
Turns out, we *were* handling A-presses on press rather than on release. We were *firing* the "mouse up" event as a proxy, but we were still firing it from code that runs when A is pressed and not when A is released.

I don't know why LockPickMenu::HandleMouseDown only works when called that way, but I don't need to know, because nothing is broken and I don't need to change anything. So basically all of this was a waste of time.