UI pref work
Sliders don't work because scrollbar user5 isn't actually the initial value: it MUST be zero; executables can set the scrollbar value by setting user5 to -999999. then to the desired position, and then back to 0, all in a single frame.

Trying to make an operator that only returns its value once. It's... proving difficult. Among other things, if the operator is dependent on multiple values, then it'll return if any of those values have resolved and not if all of them have. There's no way to check that all of the operators in the containing trait have resolved either, so an "initial only" operator isn't possible. This is a very, very difficult problem.

See, the containing trait is brought into existence with all of its operators referring to external sources... but not all of those external sources have their values yet. They'll get those values one by one as they load in, and then propagate those values to the operators that are pulling them via src. There's not really any fixed order that this happens in -- the design of the UI system doesn't consider "all values ready" a special event or even a definable one. If you wanted to initialize the trait's value to anything different, you'd need special hooks to do it at parse time, and that obviously rules out run-time computation -- which, of course, we need in order to even be able to pull UI prefs into XML -- leaving you with consts only.

Making an operator that only returns a value if its argument is non-zero, and then kills itself afterward so that it never returns a value again, is doable, but it's going to be extremely brittle. Something like <op><a /><b /></op> will return whichever of A or B resolve first -- and that really is the best we can do, at least like this.

A better workaround might be to just have the scrollbar thumbs for UI prefs always offset themselves by 1, and have the scrollbar proper undo that offset for its user7.