UI pref work
Solved the "initialize trait" problem by adding an operator that loads a pref's value exactly once and then destroys itself.

Currently, the sliders for UI prefs can be moved properly, but they don't initialize properly: they correctly obtain the pref value, move the thumb, and so on... and then *after that*, the thumb's _value_scrolled_to is somehow set to zero, without updating the traits that pull from it (i.e. without visibly being at zero). I have no idea what's causing this.

Moreover, I've realized a key design flaw in my idea: there's no way to forcibly update tiles in other menus that are pulling from a UI pref. If a tile in HUDMainMenu pulls from a UI pref for, say, scaling, and we change that pref's value from XXNOptionsMenu, there's nothing out there that can poke HUDMainMenu to get it to update.

We're gonna need to retool the pref system significantly: we need to maintain a tile in-memory that holds traits corresponding to all prefs, and we need to have markup pull from those traits directly, as one would for options in NorthernUI() or xxnLocalization(). Pref values will need to be synched to that special tile. And on top of all that, we still have to figure out why scrollbars refuse to properly initialize with AND KEEP the pref value.