UI pref work
New plan:

We synchronize all prefs to a special UI tile accessible with the xxnPrefs() selector; every pref gets an underscore-prefixed trait.

We no longer allow XML to directly set the value of a pref. Instead, scrollbars check for clicks to their various buttons and ask to modify the pref's value based on the click. Then, the scrollbars pull the current value down from xxnPrefs(). Because XML isn't responsible for maintaining pref state anymore, we don't need to faff about with "initial values" and "run-once operators" and whatnot.