UI pref work
Fixed a mistake made when implementing xxnPrefs(): we didn't actually handle "xxnPrefs()" as a selector; we just created the tile in-memory. The "!xxn tileInfo" extended console command is now also capable of dumping the xxnPrefs() tile's contents.

Fixed a parser bug involving vanilla traits with string values. Given a trait like <user0>_trait</user0>, if there existed a underscore-prefixed trait named "_trait" that had been loaded at any point earlier, then the user0 trait would be given a float value equal to the internal ID of "_trait." We now ensure that user0 gets the string it's supposed to have. To be honest, NorthernUI and MenuQue both do so much stuff to the game's parser that I'm not even sure this bug is a vanilla one.

Fixed crashes stemming from DoActionEnumeration recursion. As part of this fix, operators that work with prefs now change the pref values and then queue the updated values to be pushed back to the UI layer later; "later" here refers to "when we are no longer in any DoActionEnumeration call stack."

The system almost works perfectly. I now have to straighten out behavior for the "Reset to Defaults" button. Currently our XML only listens for that button being clicked; there's no way to react to the confirmation prompt that follows. What we need to do is: when the user confirms a reset-to-defaults, we need to set a root trait on the XXNOptionsMenu to 1 and then to 0 in the same frame; we then need to have the UI prefs use that trait. (Moreover, I want to add submenus built in pure XML; to get that working, I'll need to give every hardcoded option a trait that indicates whether the executable is allowed to reset it, and have markup set that trait to true when the option's containing submenu is open. UI prefs will also need to check whether their containing submenu is open as one of the conditions for their reset operator.)