UI pref work
UI prefs can now use all of the XML entities defined in Oblivion's spec.

XMLDocument now stores entities as std::string, so that you don't need to muck around with manually duplicating const char* instances (a concern when copying XML entities out of Oblivion's storage). This in theory also allows you to construct new entities at run-time and trust that they won't leak when the XMLDocument is destroyed.

Currently trying to build XML for a UI pref checkbox. At present, it doesn't work: the checkbox doesn't properly react when you click it, and it's not keyboard- or gamepad-accessible because Enter and the A button don't currently forward to the "checkbox" tile in the prefab.