UI pref work and XXNOptionsMenu changes
XXNOptionsMenu now has submenus implemented in XML along with hardcoded hooks to make this feasible. We can now define the submenus entirely in markup.

When you tell XXNOptionsMenu to reset options to default, and confirm that choice, it now sets a trait on its root tile to signal that this has occurred, so that UI prefs can react appropriately. Moreover, the menu only resets hardcoded options if specific tiles (the "value" for enumpickers and the container for others) have user24 set to true; this allows us to use XML to scope a reset-to-defaults to the currently-visible submenu only.

Better documention in our pause-menu option prefabs, now that we understand cyclical references in UI space a little better.

Added a submenu to XXNOptionsMenu for testing purposes and moved three of the test UI prefs into it.

Removed debug logging from most of the UI pref operators.

Uploaded Tile.h and Tile.cpp changes that I forgot to bundle in a previous commit -- these definitions are needed to allow UIPrefsManager to import all of Oblivion's XML entity definitions into its XMLDocument instance.

Miscellaneous cleanup for year-old-ish code.