Fix for an incompatibility with Ultimate Leveling
If keyboard navigation forwards clicks to a given tile by way of traits like XBUTTONX, Oblivion calls Menu::HandleMouseUp with the ID of the target tile, but a pointer to the source tile i.e. the tile that had the XBUTTONX trait. This is the underlying cause of the SleepWaitMenu softlock in the vanilla game, but it also causes MenuQue to break keyboard navigation in some menus when SetMenuEventHandler is used. This caused an incompatibility with Ultimate Leveling that prevented gamepad users from using HandleMouseUp-based gamepad buttons in BookMenu.

Fixed.