HUD scaling options
Removed all testing prefs from XXNOptionsMenu and removed all testing markup from MessageMenu. Deleted some testing files from the /data/NorthernUI/prefs/ folder.

Implemented prefs for HUD scaling, and added them to XXNOptionsMenu.

Fixed a bug in pref loading that caused prefs not to push their values to UI state (i.e. to xxnPrefs() where the UI can use them) on load. This was the result of refactoring our patches to ensure that UIPrefsManager can use Oblivion's XML entities in its pref files: we needed the manager to run only after selectors load (by which time we can guarantee that XML entities are available), but we forgot to move the push-to-UI to that time.