Improved compass scaling
Most icons in the compass now use 64x64px assets instead of 32x32px. Map markers can use 32px or 64px assets based on final rendered size; this was done because some parts of some marker icons get mangled when downscaled too hard.

The compass itself now has higher-resolution assets. The new assets are also more on-model.

In addition to rescaling the compass (which changes its thickness), you can now also adjust its width using a separate pref.

XML can now check how the HUD is being scaled by way of a trait written to the NorthernUI() tile.

If the NorthernUI() tile's XML file is missing, we now just create a blank tile in memory, allowing other patches to continue to write values to it.

Starting to upload PNG versions of the source graphics assets to the repo. (I will not be uploading the source XCF files due to their filesize, or the final DDS files due to both their filesize and their lossy compression. PNG is compact and lossy, which makes it ideal to work with here.)

Miscellaneous notes:

The icon for player-placed waypoints is the same as the icon for quest markers, using GIMP 2.8's Colorize filter; hue 183, saturation 75, lightness -35.

The icon for objective-marked doors uses the "door" map marker icon at 75% size with the bars removed, and the normal quest marker at 75% size.

The 32px map marker file is a downscaled copy of the 64px map marker file with some manual changes: the single-house icon was cleaned up, and the "E" and "C" markers on the mod-added icons were cleaned up. Map marker designs are mostly unchanged with the exception of the Obelisk of Order icon, which was redrawn from scratch.