HUD layout pref
The "pref carousel" operator has had its behavior changed, so that it can be used for both checkboxes and enumpickers. Instead of being a variant on modulo that returns the righthand term if both terms are equal, it is now a variant on modulo that always returns the righthand term in situations where modulo would return zero.

Created a prefab allowing for "enumpicker" prefs.

Removed unused math -- remnants of vanilla code -- from the compass XML.

Started work on a "HUD Layout" pref that will allow the user to show the health, magicka, and stamina bars in a vertical stack (akin to the vanilla UI). I need to revise how we position the equipped weapon and magic icons; we can't use the vanilla positioning since we conditionally show condition and charge meters for the weapon, so this is quite a challenge!