HUD layout
Finalized the different HUD layout options.

All "option" prefabs intended for use in pause/option menus now have a comment explaining how to properly handle their listindex and y traits.

Fixed an issue in the UI pref enumpicker prefab that caused it to not display the pref's value properly on load.

The xxnOpPrefCarousel operator now does nothing if the operand is less than or equal to zero, since it wouldn't make sense to do anything.

Pref values are now stored in the OBSE/Plugins folder, next to NorthernUI.ini. Just kinda seems to make more sense that way. Also, we now attempt to ignore <pref/> elements that are not direct children of the root node (though I need to test that that works).