Loading screen source XCF files
GIMP now offers the option to zlib-compress XCF files, so I've been able to cut the total file size of all loading screen XCF files by about a third. These XCF files are what was used to generate the loading screens: I would load the original screenshot, immediately dupe it to a new layer, and use the smudge tool to anti-alias everything by hand. In a few screenshots, I also added some extra effects, such as noise on distant terrain or bloom effects when the sun was partially blocked by mountains. All of that work is preserved in the XCFs.

The title screen got the most extensive work, with the foreground scenery darkened and the color balance shifted.